Agents/AutoGen/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
autogen-agentchat==0.4.9.2
autogen-core==0.4.9.2
autogen-ext==0.4.9.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
jsonref==1.1.0
openai==1.68.2
opentelemetry-api==1.31.1
pillow==11.1.0
protobuf==5.29.4
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tavily-python==0.5.1
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
